chore: move `PartialWitnessGenerator` test to `pwg.rs`
