+++ acvm/src/pwg.rs
@@ -180,3 +180,214 @@ fn insert_value(
 
     Ok(())
 }
+
+#[cfg(test)]
+mod test {
+    use std::collections::BTreeMap;
+
+    use acir::{
+        circuit::{
+            directives::Directive,
+            opcodes::{FunctionInput, OracleData},
+            Opcode,
+        },
+        native_types::{Expression, Witness, WitnessMap},
+        FieldElement,
+    };
+
+    use crate::{
+        pwg::{self, block::Blocks, OpcodeResolution, PartialWitnessGeneratorStatus},
+        OpcodeResolutionError, PartialWitnessGenerator,
+    };
+
+    struct StubbedPwg;
+
+    impl PartialWitnessGenerator for StubbedPwg {
+        fn aes(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _inputs: &[FunctionInput],
+            _outputs: &[Witness],
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn and(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _lhs: &FunctionInput,
+            _rhs: &FunctionInput,
+            _output: &Witness,
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn xor(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _lhs: &FunctionInput,
+            _rhs: &FunctionInput,
+            _output: &Witness,
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn range(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _input: &FunctionInput,
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn sha256(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _inputs: &[FunctionInput],
+            _outputs: &[Witness],
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn blake2s(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _inputs: &[FunctionInput],
+            _outputs: &[Witness],
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn compute_merkle_root(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _leaf: &FunctionInput,
+            _index: &FunctionInput,
+            _hash_path: &[FunctionInput],
+            _output: &Witness,
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn schnorr_verify(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _public_key_x: &FunctionInput,
+            _public_key_y: &FunctionInput,
+            _signature: &[FunctionInput],
+            _message: &[FunctionInput],
+            _output: &Witness,
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn pedersen(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _inputs: &[FunctionInput],
+            _outputs: &[Witness],
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn hash_to_field_128_security(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _inputs: &[FunctionInput],
+            _output: &Witness,
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn ecdsa_secp256k1(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _public_key_x: &[FunctionInput],
+            _public_key_y: &[FunctionInput],
+            _signature: &[FunctionInput],
+            _message: &[FunctionInput],
+            _output: &Witness,
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn fixed_base_scalar_mul(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _input: &FunctionInput,
+            _outputs: &[Witness],
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+        fn keccak256(
+            &self,
+            _initial_witness: &mut WitnessMap,
+            _inputs: &[FunctionInput],
+            _outputs: &[Witness],
+        ) -> Result<OpcodeResolution, OpcodeResolutionError> {
+            panic!("Path not trodden by this test")
+        }
+    }
+
+    #[test]
+    fn inversion_oracle_equivalence() {
+        // Opcodes below describe the following:
+        // fn main(x : Field, y : pub Field) {
+        //     let z = x + y;
+        //     constrain 1/z == Oracle("inverse", x + y);
+        // }
+        let fe_0 = FieldElement::zero();
+        let fe_1 = FieldElement::one();
+        let w_x = Witness(1);
+        let w_y = Witness(2);
+        let w_oracle = Witness(3);
+        let w_z = Witness(4);
+        let w_z_inverse = Witness(5);
+        let opcodes = vec![
+            Opcode::Oracle(OracleData {
+                name: "invert".into(),
+                inputs: vec![Expression {
+                    mul_terms: vec![],
+                    linear_combinations: vec![(fe_1, w_x), (fe_1, w_y)],
+                    q_c: fe_0,
+                }],
+                input_values: vec![],
+                outputs: vec![w_oracle],
+                output_values: vec![],
+            }),
+            Opcode::Arithmetic(Expression {
+                mul_terms: vec![],
+                linear_combinations: vec![(fe_1, w_x), (fe_1, w_y), (-fe_1, w_z)],
+                q_c: fe_0,
+            }),
+            Opcode::Directive(Directive::Invert { x: w_z, result: w_z_inverse }),
+            Opcode::Arithmetic(Expression {
+                mul_terms: vec![(fe_1, w_z, w_z_inverse)],
+                linear_combinations: vec![],
+                q_c: -fe_1,
+            }),
+            Opcode::Arithmetic(Expression {
+                mul_terms: vec![],
+                linear_combinations: vec![(-fe_1, w_oracle), (fe_1, w_z_inverse)],
+                q_c: fe_0,
+            }),
+        ];
+
+        let backend = StubbedPwg;
+
+        let mut witness_assignments = BTreeMap::from([
+            (Witness(1), FieldElement::from(2u128)),
+            (Witness(2), FieldElement::from(3u128)),
+        ])
+        .into();
+        let mut blocks = Blocks::default();
+        let solver_status = pwg::solve(&backend, &mut witness_assignments, &mut blocks, opcodes)
+            .expect("should stall on oracle");
+        let PartialWitnessGeneratorStatus::RequiresOracleData { mut required_oracle_data, unsolved_opcodes } = solver_status else {
+            panic!("Should require oracle data")
+        };
+        assert!(unsolved_opcodes.is_empty(), "oracle should be removed");
+        assert_eq!(required_oracle_data.len(), 1, "should have an oracle request");
+        let mut oracle_data = required_oracle_data.remove(0);
+
+        assert_eq!(oracle_data.input_values.len(), 1, "Should have solved a single input");
+
+        // Filling data request and continue solving
+        oracle_data.output_values = vec![oracle_data.input_values.last().unwrap().inverse()];
+        let mut next_opcodes_for_solving = vec![Opcode::Oracle(oracle_data)];
+        next_opcodes_for_solving.extend_from_slice(&unsolved_opcodes[..]);
+        let solver_status =
+            pwg::solve(&backend, &mut witness_assignments, &mut blocks, next_opcodes_for_solving)
+                .expect("should be solvable");
+        assert_eq!(solver_status, PartialWitnessGeneratorStatus::Solved, "should be fully solved");
+    }
+}
